+++ src/main/kotlin/org/openartifact/artifact/graphics/window/WindowConfig.kt
@@ -1,3 +1,3 @@
 package org.openartifact.artifact.graphics.window
 
-data class WindowConfig(val width: Int, val height: Int, val title : String)
+data class WindowConfig(val width: Float, val height: Float, val title : String)
